+++ client/src/components/FeaturedPost.js
@@ -7,32 +7,30 @@ import CardActionArea from '@mui/material/CardActionArea';
 import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 
-function FeaturedPost(props) {
+function FeaturedPost (props) {
   const { post } = props;
 
   return (
-    <Grid item xs={12} md={6} mb={3}>
-      <CardActionArea component="a" href="#">
-        <Card sx={{ display: 'flex' }}>
-          <CardContent sx={{ flex: 1 }}>
-            <Typography component="h2" variant="h5">
-              {post.title}
-            </Typography>
-            <Typography variant="subtitle1" color="text.secondary">
-              {post.date}
-            </Typography>
-            <Typography variant="subtitle1" paragraph>
-              {post.description}
-            </Typography>
-          </CardContent>
-          <CardMedia
-            component="img"
-            sx={{ width: 160, display: { xs: 'none', sm: 'block' } }}
-            image={post.image}
-            alt={post.imageLabel}
-          />
-        </Card>
-      </CardActionArea>
+    <Grid item xs={12} mb={3}>
+      <Card sx={{ display: 'flex' }}>
+        <CardContent sx={{ flex: 1 }}>
+          <Typography component='h2' variant='h5'>
+            {post.title}
+          </Typography>
+          <Typography variant='subtitle1' color='text.secondary'>
+            {post.date}
+          </Typography>
+          <Typography variant='subtitle1' paragraph>
+            {post.description}
+          </Typography>
+        </CardContent>
+        <CardMedia
+          component='img'
+          sx={{ width: 160, display: { xs: 'none', sm: 'block' } }}
+          image={post.image}
+          alt={post.imageLabel}
+        />
+      </Card>
     </Grid>
   );
 }
@@ -43,8 +41,8 @@ FeaturedPost.propTypes = {
     description: PropTypes.string.isRequired,
     image: PropTypes.string.isRequired,
     imageLabel: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-  }).isRequired,
+    title: PropTypes.string.isRequired
+  }).isRequired
 };
 
 export default FeaturedPost;
