+++ client/src/components/Map.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { ReactComponent as SouthKorea } from '../assets/svg/south-korea2.svg';
 import ReactToolTip from 'react-tooltip';
+import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
 
 // web3
 const Web3 = require('web3');
@@ -48,8 +50,8 @@ const Map = () => {
         return {
           display: 'flex',
           justifyContent: 'center',
-          marginTop: '2em',
-          marginBottom: '2em',
+          paddingTop: '3em',
+          paddingBottom: '6em',
           maxWidth: theme.breakpoints.values.md
         };
       }}
