+++ client/src/pages/Main.js
@@ -1,46 +1,17 @@
-import Box from '@mui/material/Box';
-import Grid from '@mui/material/Grid';
 import Container from '@mui/material/Container';
-import FeaturedPost from 'components/FeaturedPost';
-import Map from '../components/Map';
-import ReactTooltip from 'react-tooltip';
+import Divider from '@mui/material/Divider';
 
-const featuredPosts = [
-  {
-    title: 'Posco International to invest 3.8 trillion won in energy business',
-    date: 'Dec 6',
-    description:
-      'Posco International, which is set for a merger with Posco Energy, will invest 3.8 trillion won ($2.7 billion) over the next three years in its energy businesses, the trading company said Friday. Posco International’s merger with Posco Energy, a liquefied natural gas (LNG) plant operator, was approved at an extraordinary shareholders’ meeting held at the company’s headquarters in Songdo, Incheon, Friday.',
-    image: 'https://source.unsplash.com/random',
-    imageLabel: 'Image Text'
-  },
-  {
-    title: 'POSCO International Aims to Become Global Energy Giant',
-    date: 'Dec 5',
-    description:
-      'According to the Data Analysis, Retrieval and Transfer System (DART) of the Financial Supervisory Service on Oct. 30, POSCO International achieved solid results in the third quarter based on a diversified business portfolio amid a global economic downturn. Its sales came in at 9.41.2 trillion won in the third quarter, down 1.2 percent from the same period last year. Its operating profit swelled 32.7 percent to 197 billion won and its net profit 63.5 percent to 112.8 billion won.',
-    image: 'https://source.unsplash.com/random',
-    imageLabel: 'Image Text'
-  }
-];
+import Map from 'components/Map';
+import ReactTooltip from 'react-tooltip';
+import News from 'components/News';
 
 function Main () {
   return (
     <Container maxWidth='lg'>
-      <Box
-        component='main' sx={{
-          display: 'flex',
-          height: '100%'
-        }}
-      >
-        <Grid container spacing={4}>
-          {featuredPosts.map((post) => (
-            <FeaturedPost key={post.title} post={post} />
-          ))}
-        </Grid>
-        <Map className='mapbox' marginBottom='20px' title='Shepherd: Current Supplychain' />
-        <ReactTooltip place='bottom' effect='solid' type='warning' />
-      </Box>
+      <Map className='mapbox' marginBottom='20px' title='Shepherd: Current Supplychain' />
+      <ReactTooltip place='bottom' effect='solid' type='warning' />
+      <Divider />
+      <News sx={{ pt: 4, pb: 6 }} />
     </Container>
   );
 }
